qemubios: allocate new pages when loading kernel
We're having repeated issues with the ELF file not containing zeros
between filesz and memsz.
This is technically valid, but also highly annoying and forces us to
make a copy of the pages.

If we have to copy the pages anyway, just allocate new pages for
all segments and copy the data over. Much easier that way.