IDL: derive API IDs from hash of file
This is to help ensure specifications are not modified without things
breaking in a very obvious manner.

There is in theory a risk of ID collisions due to the birthday problem,
but I do not anticipate this to be a problem in practice. There are no
security implications either, if one such collisions does happen to
occur in the wild.