kernel: keep track of current thread's TCB
This is necessary to correctly save and restore thread state.

We will not use a global. This is because, after much deliberation, it
is decided to allow multiple hardware threads (HARTs) to share a single
scheduler.

We currently use the FS register, though ideally we should have a proper
per-hart state structure and free up the FS register for user
applications.