IDL/gen_rust.py: rework code generation
The old "manually count registers" was retarded, to be frank.
It is much too difficult to match ABI that way.
Returning more than "3 registers" was especially problematic. And the
reason to rework it now, as a soon-to-be-added API wasn't compatible
with it.

The new generation follows a much simpler rule: if it's a record,
unpack it before passing as arguments.
Sum types are somewhat tricky, but we mandate that field types align
perfectly, so just look at the first variant and recurse until
non-sum-type.

Return types remain untouched.

PS: I'm confident handling of sum types in argument position is
broken... but it should barf when we get there so whatever.