kernel: only unlock ThreadManager after saving thread state
An issue, in some upcoming work, is the following situation on thread
switch:

- thread B is dequeued by thread A
- thread A enqueues itself
- thread B is resumed

The issue: we need to release the thread manager *before* resuming
thread B but *after* saving thread A. If we don't do this, thread A may
get dequeud by another hart before the state is saved.
(this is not an issue without SMP thanks to critical sections)

The only location where we can do this properly is inside the assembly
block.