Events are dead. Long live events.
After getting mentally stuck on a different branch I've come to realize
the event system is not only flawed but contrary to one of the primary
goals of this OS.

The first issue: events break the threaded model by requiring to send
messages and wait instead. This can be abstracted away of course, but
then what is the point? Going threaded all the way simplifies things and
should be more efficient too.

Secondly, the original reason for switching to the event model was to
be able to *adjust* timers while also waiting on queue events in the
virtio-net driver. However, there is a much simpler solution: timer
objects! With timer objects one thread can wait for queue events,
another on a timer and any other thread can manipulate the timer at
will.