Add critical sections to thread API
As we removed the events API we overlooked something:
we need to support external thread queues while also avoiding
a race condition between unlocking the queues and parking.

Perhaps it would be worth having a per-thread lock?
Acquire the lock when running and release it when scheduling.
It would automagically avoid race conditions even if a thread gets
unscheduled before parking.

That would make this commit redundant, but keep it nonetheless as
documentation.