IDl: simplify allocator
The primary change is the default assumption that
physical address == virtual address for situations where this is
relevant.

This should greatly simplify things, but can also bite us very hard
later on:

- We now have a single allocator API that is useable for
  drivers+services in kernel space as well as user-space processes.
- No need to bother with 64- vs 32-bit differences (or 16 bit...)

On the other hand:

- We have to keep the identity map in the lower half. Unconvential, but
  it just might work out for this particular design...
- IOMMU or bounce buffers is pretty much required if ever need to have
  32-bit addresses on a 64-bit platform.

The latter situation can also be addressed with an explicit 32-bit allocator
alongside this allocator + bounce buffers.
But such a thing does not need to be directly integrated.