kernel: add priority level for idle threads
We were stubborn at first because we wanted to avoid a panic branch in
PriorityQueue::dequeue, but not hqving a queue for idle threads makes
things more complicated than necessary.

In particular, we were suffering from the idle thread getting enqueued
when it should not be. Fixable, but at the cost of branches. And what
would then the point of avoiding the panic branch be?
At least the panic branch is highly predictable and hence practically
zero-cost. If deemed necessary, we can even use an unreachable intrinsic.