kernel: don't log to framebuffer anymore
We overlooked one major pitfall: stack overflow during logging.

If this happens, it is possible the lock to the framebuffer won't get
released. Then we try to log during the page fault handler and we
deadlock. Ouch!

It is *definitely* worth getting a separate stack for some of these
critical locks, but that can be done later.