lib/virtio{,-net}: fix packet capacity to 1514, separate header
For receiving packets it makes little sense to go below this.
For sending packets it's whatever.

Separate the virtio-net header from the packet so we can take buffers
from wherever and attach the necessary header to it without copying,
or somehow making other processes prescient.

This is an API break, but to avoid creating a huge commit as a bunch
more changes are staged, allow it for this commit.